feat: add check in delete in services

+++ src/main/java/ru/kurbanov/application/services/CarServiceImpl.java
@@ -162,6 +162,10 @@ public class CarServiceImpl implements CarService {
 
     @Override
     public void removeCar(UUID carId) {
+        if (!carRepository.existsById(carId)) {
+            throw new EntityNotFoundException("Detail not found: " + carId);
+        }
+
         carRepository.deleteById(carId);
     }
